Rename `funcMisc.py` -> `functions.py`.
