+++ src/communications/ig_rest_api.py
@@ -9,7 +9,7 @@ import traceback
 from collections import OrderedDict
 from copy import deepcopy
 
-from src.gui.funcMisc import read_config, write_config
+from src.gui.functions import read_config, write_config
 
 
 class APIError(Exception):
