+++ src/exports/excel.py
@@ -4,7 +4,7 @@ import re
 from collections import OrderedDict
 import numpy as np
 
-from src.gui.funcMisc import read_config
+from src.gui.functions import read_config
 
 RE_TEXT_BETWEEN_TAGS = re.compile(r">(.*?)<")
