+++ src/gui/widgets.py
@@ -8,7 +8,8 @@ from PyQt5 import QtGui, QtWidgets
 import re
 import datetime
 
-from src.gui.funcMisc import read_config
+from src.gui.functions import read_config
+
 RE_LABEL = re.compile(r"(.*?[A-z]): ")
 RE_TAG = re.compile(r"<(.*?)>")
